refactor: Modify header search layout & error
