+++ src/components/layout/footer/index.tsx
@@ -3,7 +3,6 @@ import { Box } from "./FooterStyles";
 const Footer = () => {
   return (
     <Box>
-      {/* <Logo src="/logo.svg" alt="" /> */}
       <p className="text-sm text-gray-500 sm:mt-0 dark:text-gray-300">
         © 2024-yju-capston-team3 Copyright 2023. All Rights Reserved.
       </p>
