+++ src/components/map/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from "react";
+import { useState, useEffect, useMemo } from "react";
 import { GoogleMap, LoadScript, MarkerF } from "@react-google-maps/api";
 import { Wrapper, Styles } from "./MapStyles";
 import { getAllPoints } from "../../services/location.service";
